moving generic functions to seperate file
